Fix some tests

+++ orePlayCommon/app/db/impl/query/SharedQueries.scala
@@ -9,6 +9,6 @@ object SharedQueries {
 
   val refreshHomeView: doobie.Update0 = sql"REFRESH MATERIALIZED VIEW home_projects".update
 
-  def watcherStartProject(id: DbRef[Project]): doobie.Query0[(Int, Int)] =
-    sql"""SELECT p.stars, p.watchers FROM home_projects p WHERE p.id = $id""".query[(Int, Int)]
+  def watcherStartProject(id: DbRef[Project]): doobie.Query0[(Long, Long)] =
+    sql"""SELECT p.stars, p.watchers FROM home_projects p WHERE p.id = $id""".query[(Long, Long)]
 }
